chore: refactor quiz component
