+++ README.md
@@ -27,7 +27,7 @@ A web application that allows users to upload a PDF file and generate a multiple
 - [Tailwind CSS](https://tailwindcss.com/) – Utility-first CSS framework
 - [Shadcn/UI](https://ui.shadcn.com/) – Accessible and beautiful UI components
 - [Zustand](https://github.com/pmndrs/zustand) – Lightweight global state management
-- [PDF Parse New](https://www.npmjs.com/package/pdf-parse-new) – For extracting text from PDF files
+- [pdf-parse-new](https://www.npmjs.com/package/pdf-parse-new) – For extracting text from PDF files
 - [OpenAI API](https://platform.openai.com/docs) – For generating quiz questions using AI
 - [Sonner](https://sonner.emilkowal.ski/) – Toast notification system
 
@@ -63,7 +63,7 @@ MAX_FILE_SIZE=your-max-file-size
 MAX_API_CALL_TIMEOUT=your-max-api-call-timeout
 ```
 
-- Important: Make sure not to commit this file.
+> Important: Make sure not to commit this file.
 
 ---
 
@@ -80,11 +80,11 @@ Then open your browser at http://localhost:3000.
 ```bash
 src/
 ├──  app/
-├──  components/
 ├──  features/
 │  ├──  upload-pdf/
 │  ├──  quiz/
 │  ├──  quiz-history/
+├──  components/
 ├──  hooks/
 ├──  lib/
 │  ├──  schema/
@@ -111,7 +111,51 @@ src/
 - State is kept minimal and persisted in localStorage via Zustand
 - Large PDFs are processed in chunks to avoid performance issues and API limits
 - Quiz generation is asynchronous and error-resilient
+- Real-time feedback with sonner toast notifications
 - All components use functional React and modern hooks
+
+---
+
+## User Guide
+
+## Step 1: Upload a PDF File
+
+1. On the home page, click the **"Choose File"** button.
+2. Select a PDF file from your computer.
+3. The file name will appear once selected.
+
+> **Note:** The app supports large PDF files and processes them in chunks for optimal performance.
+
+## Step 2: Generate Quiz
+
+1. After uploading the file, click the **"Generate Quiz"** button.
+2. The app will extract the text from the PDF and send it to the OpenAI API to generate quiz questions.
+3. A loading indicator or toast will show the process status.
+
+> If something goes wrong (e.g., file issues or API errors), a message will be shown using **Sonner** toast notifications.
+
+## Step 3: Answer the Quiz
+
+1. The generated multiple-choice quiz will be displayed.
+2. Select your answer for each question.
+3. Once you've answered all questions, click the **"Finish Quiz"** button.
+
+## Step 4: View Results
+
+- After submission, you'll see:
+  - Your **score**
+  - A summary showing correct vs. incorrect answers
+  - Which answer was the right one for each question
+
+## Step 5: Review Quiz History
+
+1. On the home screen or sidebar, go to the **Quiz History** section.
+2. You will see a list of previously generated quizzes (stored using **Zustand** in `localStorage`).
+
+For each quiz:
+
+- **Retake** – Reattempt the quiz from scratch
+- **Delete** – Remove it permanently from history
 
 ---
 
@@ -121,3 +165,4 @@ src/
 - Minimal error handling for scanned/invalid PDFs
 - No handling for OpenAI rate limits or token overflow
 - No quiz export or print feature
+- Some parts of the mobile experience can be improved
