+++ src/app/api/generate-quiz/route.ts
@@ -8,9 +8,6 @@ import { generateQuizFromPdf } from "@/lib/utils/generate-quiz-from-pdf";
  * @returns {Promise<Response>} - A JSON response containing the generated quiz or an error message.
  *
  * @throws {Error} If the OpenAI API key is missing or if any step in the quiz generation fails.
- *
- * @example
- * // Request should be a multipart/form-data with a 'pdf' field containing the PDF file.
  */
 export async function POST(request: NextRequest) {
   try {
