+++ src/hooks/use-quiz-question.ts
@@ -1,0 +1,62 @@
+import { QuizData } from "@/types/quiz.type";
+import { useState } from "react";
+
+export function useQuizQuestion(
+  quiz: QuizData,
+  onComplete: (score: number) => void
+) {
+  const [currentQuestion, setCurrentQuestion] = useState(0);
+  const [selectedAnswers, setSelectedAnswers] = useState<number[]>([]);
+  const [showResults, setShowResults] = useState(false);
+  const [selectedAnswer, setSelectedAnswer] = useState<number | null>(null);
+
+  const handleAnswerSelect = (answerIndex: number) => {
+    setSelectedAnswer(answerIndex);
+  };
+
+  const handleNextQuestion = () => {
+    if (selectedAnswer === null) return;
+
+    const newAnswers = [...selectedAnswers];
+    newAnswers[currentQuestion] = selectedAnswer;
+    setSelectedAnswers(newAnswers);
+
+    if (currentQuestion < quiz.questions.length - 1) {
+      setCurrentQuestion(currentQuestion + 1);
+      setSelectedAnswer(null);
+    } else {
+      const finalScore = calculateScore(newAnswers);
+      onComplete(finalScore);
+      setShowResults(true);
+    }
+  };
+
+  const handlePreviousQuestion = () => {
+    if (currentQuestion > 0) {
+      setCurrentQuestion(currentQuestion - 1);
+      setSelectedAnswer(selectedAnswers[currentQuestion - 1] ?? null);
+    }
+  };
+
+  const calculateScore = (answers: number[] = selectedAnswers) => {
+    return answers.reduce((score, answer, index) => {
+      return score + (answer === quiz.questions[index].correctAnswer ? 1 : 0);
+    }, 0);
+  };
+
+  const question = quiz.questions[currentQuestion];
+  const progress = ((currentQuestion + 1) / quiz.questions.length) * 100;
+
+  return {
+    question,
+    progress,
+    currentQuestion,
+    selectedAnswer,
+    selectedAnswers,
+    showResults,
+    handlePreviousQuestion,
+    handleNextQuestion,
+    handleAnswerSelect,
+    calculateScore,
+  };
+}
